Add files via upload
I made this for fun. It just uses C++ to print the * character in such a way that you get a circle that is not hollow. You will provide the radius yourself. If someone can come up with a better code or improve it I will be highly interested in your efforts and delighted to learn.